Fixed docker build.
Updated the docker file according to the new install procedure. Because python 3.11 is unsupported, the image has been switched from "alpine:edge" to "python:3.10-alpine".
Added docker_fix_xdis.py which automatically applies the hotfix described in INSTALLATION.md.
Added unnpk to the cargo-build section and the required dependencies to build it.
Updated the README.md file with instructions on how to correctly build the docker image.
Removed unnpk from .gitmodules because it does not work, it is required to manually clone it (updated the installation instructions accordingly).